requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.5
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1